Fixes #430.
This is a simpler fix for the time skew portion of the
consecutive recording problem. Part of the problem I 
was trying to solve was already solved by David Engel.
MythTV should no longer try to schedule something for 
recording that is already recording on a slave backend.
So the GetRecording() request is not needed in the 
scheduler. We can instead simply cancel any non-LiveTV
recording if the master backend requests a new recording.



git-svn-id: http://svn.mythtv.org/trunk@7433 4b91aca1-2a1f-0410-a73a-a78f2df169e8