This makes the signal monitoring UI a little nicer, and cleans up some bounds issues with the dvbsignalmonitor.
The "F7" antenna mode does not turn on table monitoring
anymore, making it much more useful for adjusting an
antenna. And I also fixed a bug introduced in 7421,
antenna adjustment requires signal notifications,
not just live tv...

The signal monitoring UI will fall back to S/N ratio
if the signal quality value is not useful, and fall
back further to BER if the S/N ratio is not good.
Assuming those other values are set and are useful.

For the S/N ratio I did a little hack that should
allow drivers that improperly set this as an unsigned
value to work. It does assume that this value is raw
and not in decibels. The DVB API does not specify
that this should be a raw value, but this makes the
most sense since there is no fixed point defined and
it is harder to do a log base 10 in the kernel than
in user space. If someone has a card that returns
this value in dB please tell me and we can work out
some hack that makes it work.



git-svn-id: http://svn.mythtv.org/trunk@7449 4b91aca1-2a1f-0410-a73a-a78f2df169e8