Makes MythTV a little smarter with respect to channel separator.
This adds a static function:
   SourceUtil::GetChannelSeparator(uint sourceid)

It returns the channel separator already being used
by the greatest number of existing channels in on that
source. So it you can default to using when scanning,
and use it when inserting channels from datadirect.

So if you choose to use "-" or "" or "0" as the channel
separator when you do a scan, datadirect will use that
character instead of always using "_" when inserting
channels.

This also adds "." as a valid channel separator for the 
smart keys channel entry in tv_play.cpp



git-svn-id: http://svn.mythtv.org/trunk@7468 4b91aca1-2a1f-0410-a73a-a78f2df169e8