Replaces two questions on mythtv-setup startup with "delete all" options on the capture cards and sources configuration screens.
You will now only see a pre-setup dialog if the backend is
running and that dialog now gives you the option of exiting
the setup program or continuing. This avoids the long delay
in starting up the themed menus for the setup program if you
prefer to exit and shutdown the running backend.



git-svn-id: http://svn.mythtv.org/trunk@7470 4b91aca1-2a1f-0410-a73a-a78f2df169e8