Three mythtv-setup improvements on "input connections" pane:   * Changes "starting channel" text entry to an editable combobox     containing the current value and the channels on the connected source.   * Adds a "Scan for Channel" button above the "starting channel" entry,     so that you can add the channels if you don't have them.   * Adds a "Fetch channels from listings source" button, also above     "starting channel" entry, so that you can add the channels if     you don't have them. This button is not so nice in that it takes     a long time to run because it is actually getting today's listings     on all the sources, not just the current source.
If there were no channels before you pressed one of the channel
adding buttons, the starting channel is set to the first channel
in the "starting channel" ComboBox after it is re-loaded.

I also removed a debbugging cerr<< I accidentally left in setup/main.cpp,
and cleaned up the datadirect dialog a little bit, and clarified some
text and fixed typos in some of the text.



git-svn-id: http://svn.mythtv.org/trunk@7472 4b91aca1-2a1f-0410-a73a-a78f2df169e8