Stops playbackbox from blocking on preview generation.
If the preview doesn't exist we create a thread to create
one and use a blank preview until the non-blank one has
been generated. Once the preview is generated a redraw 
event is scheduled and we get to see the preview.

The point is that if you want to scroll down to the fifth
program in a list and play it, you don't have to wait for
all five previews to be generated first...



git-svn-id: http://svn.mythtv.org/trunk@7481 4b91aca1-2a1f-0410-a73a-a78f2df169e8