Makes process request thread pool expand if a timeout is reached and makes the initial size a define.
This means you can add as many frontends as you want and performance
will only degrade when the master backend server can no longer keep up,
not due to this arbritary process request thread count limit in MythTV.

Currently the blocking timeout before the pool is expanded is 10 ms.
This is set by a define near the top of mainserver.cpp.



git-svn-id: http://svn.mythtv.org/trunk@7495 4b91aca1-2a1f-0410-a73a-a78f2df169e8