Closes #373, by applying patch + some editing.
Parts of this patch had already made it into SVN, plus I've
made this multi-threaded software encoding default to being
disabled. Those that want to experiment with it can pass 
"--enable-ffmpeg-pthreads" to ./configure. You then need to
"make clean" before recompiling.

I haven't added this to the --help text for ./configure for now.



git-svn-id: http://svn.mythtv.org/trunk@7507 4b91aca1-2a1f-0410-a73a-a78f2df169e8