A little adjustment for NVP::AVSync()
If we are using hardware decoding, display the frame anyway, or
at least a field of it when using bob-deinterlacing. But then don't
wait for the AVSync interval. This gets us back into sync within
a few frames but doesn't waste the already done video decoding.

This could cause tearing when paired with a really fast video card
and a low frame rate display. But if we are falling behind with
video decoding, a really fast video card probably isn't a problem.



git-svn-id: http://svn.mythtv.org/trunk@7520 4b91aca1-2a1f-0410-a73a-a78f2df169e8