Fixes #478, by wrapping dvd headers in extern "C" macros.
This appears to be due to systems that update to glibc 2.3.5 without
putting using dvdread headers that have put the inttypes.h include
of ifo_types.h inside a extern "C" wrapper.

This is really a distribution problem with the dvdread package, but 
it looks like it is easy enough to work around this bug.



git-svn-id: http://svn.mythtv.org/trunk@7525 4b91aca1-2a1f-0410-a73a-a78f2df169e8