Adds hasV4L2() and probeV4LInputs() to CardUtil. Uses hasV4L2() in channel.cpp Converts center frequencies to visual carrier frequencies for ATSC Channel instances (this lets you connect a DVB input to a HDTV card without being off by 1750000 Hz when you tune. Simplifies VideoDevice::fillDVBInputDiseqc(), which is the DVB version of CardUtil::probeV4LInputs().
If a V4L card doesn't return any input, but otherwise works,
assume that it has one input called "Television". This allows
version 2.0 of pcHDTV V4L drivers for the HD-3000 to be setup
without DB hacks.

Merge DVBDefaultInput() and TunerCardInput().

This could use more cleanup, but this is a start.



git-svn-id: http://svn.mythtv.org/trunk@7541 4b91aca1-2a1f-0410-a73a-a78f2df169e8