References #494.
This moves reschedule out of eithelper into eitscanner.
It also only reschedule when we have new info and either
we change channels or haven't seen new info in over a
minute.

If we wanted to be even kinder to the DB we could also
track when these EIT's are for and only call reschedule
if they are for the next 3 hours, or we haven't rescheduled
in over 3 hours...



git-svn-id: http://svn.mythtv.org/trunk@7558 4b91aca1-2a1f-0410-a73a-a78f2df169e8