References #512.
This implements a basic chromakey OSD. 

For it to be enabled, you must have this in your 
xorg.conf "Devices" section:
    Option "XvmcUsesTextures" "false" # use 3d engine for overlays
And you have to insert a "UseChromaKeyOSD","1",hostname setting in
the settings table.

This currently makes no claim to efficiency, this clears the a
buffer to the chromakey color w/borders, and then plops the OSD
on top of it; creating an RGB OSDSurface or creating a specialized,
chromakey painting, OSD blend function would probably be much
faster. What this does make a claim to is not being too disruptive
to the rest of the code.



git-svn-id: http://svn.mythtv.org/trunk@7569 4b91aca1-2a1f-0410-a73a-a78f2df169e8