References #523.
This commit adds printing of the FourCC codes of the XVideo surfaces
when "-v playback" is enabled. It als adds the IYUV planar format to
the list of accepted formats; IYUV is bit equivalent to the I420 
format, so we just treat it as if we had seen a I420 id.

mark-clist AT npsl.co.uk, can you attach the FourCC codes it
prints out for your ATI card?

Look for "XVideo Format" in the output..



git-svn-id: http://svn.mythtv.org/trunk@7570 4b91aca1-2a1f-0410-a73a-a78f2df169e8