Add new mythwelcome feature.
Probably mostly of interest to people who have a 
dedicated frontend/backend machine that you don't
want to leave on 24/7. It provides a themed status
screen that can be shown when myth has automatically 
started the system to record something or its waiting
for the backend to become idle and shut things down. 

It shows a status message, what is currently recording, and
what the next scheduled recording is. Also includes a button 
to easily start the frontend. 

The companion program mythshutdown can among other things 
prevent the backend prematurely shutting down the system if 
it is busy commercial flagging, downloading EPG data or 
transcoding etc 

The README file in programs/mythwelcome contains
a lot more details on what it can do and how to set
things up.

Closes #407


git-svn-id: http://svn.mythtv.org/trunk@7571 4b91aca1-2a1f-0410-a73a-a78f2df169e8