References #466.
This cleans up the function where the crash actually a
happened a bit. But I don't know what actually caused
the crash, so this just makes the backtrace look better.

A new backtrace is needed of course.

But it would also be useful to know what edit keys were actually
hit and what the aspect ratio of the video and screen were, and
if there was a letterbox mode in effect. And what the resolution
of the video and screen were.



git-svn-id: http://svn.mythtv.org/trunk@7583 4b91aca1-2a1f-0410-a73a-a78f2df169e8