References #427.
This makes Channel::SwitchToInput() set the video format based on
the channel video format when it is asked to set the channel rather
being told to just switch to the input.

I think the problem may have been that SwitchToInput() was failing
because the default format was invalid for the input, so it never
got to the point of setting the correct format when it set the
channel.



git-svn-id: http://svn.mythtv.org/trunk@7597 4b91aca1-2a1f-0410-a73a-a78f2df169e8