Closes #507, by giving option of a SLOW scan.
Basically the problem is that that the DVB drivers used by
the "Leadtek LR6650" don't report an FE_LOCK in a reasonable
time, even after they are delivering data to MythTV they are
telling MythTV that they don't see any data out there...
This is a serious driver bug. But my experience with DVB 
suggests that serious driver bugs are commonplace so I've added
the option of allowing the channel scan to ignore the FE_LOCK
status timeout altogether. This, of course, makes the scan take
much longer, but it's probably preferable to not getting any
channels from the scan.

Users that require this option may experience problems in
MythTV as well, but MythTV tuning is a little more forgiving
of broken drivers than the MythTV channel scan so YMMV.



git-svn-id: http://svn.mythtv.org/trunk@7654 4b91aca1-2a1f-0410-a73a-a78f2df169e8