Makes the record of livetv a little smarter.
If a recording was already started but then canceled, this 
reactives it when you request a recording instead of just
toggling the recording status, which won't do anything.

Also, if we are recording something and the user toggles the 
record, it pops up a message to go to the program guide to cancel
a recording. The logic here is that if you hit the record button
on something that is already recording and that you are watching,
it may be because you don't know it is already being recorded.
I think it is much less likely that you actually want to cancel
the recording...



git-svn-id: http://svn.mythtv.org/trunk@7656 4b91aca1-2a1f-0410-a73a-a78f2df169e8