Cleans up DTVRecorder::FindKeyframes() a little.
This uses the PESStreamID enum to look up the streamID's
by name rather than hex value.

This also prints out a message if MythTV finds a keyframe
that is not in the first TS packet in the series of packets
containing a PES packet. Allowing for this properly would
require quite a bit of buffering (almost a full TS->PES
transformation). This is technically allowed by the basic
MPEG standards, but standard compliant MPEG TS hardware muxing
devices are not required to work on such streams. I believe
this makes these types of packets non-existant in the wild.
If I'm wrong a message telling the user to mail the mythtv-dev
mailing list will be printed out... MythTV won't crash
or anything in such a case, it will simply construct a bad
keyframe map.



git-svn-id: http://svn.mythtv.org/trunk@7675 4b91aca1-2a1f-0410-a73a-a78f2df169e8