Fixes #513. Fixes some text input issues in LiveTV mode.
The problem is that with smart channel change enbled you
can't enter a value for the Teletext page or the arbitrary
skip that has a partial match with a channel on the current
source listings source.

The solution for teltext/cc is to make users enter the Teletext
key before entering a teletext page number, and then either press
the Teletext key again or press select. If the user wants to
perform a simple toggle they simply have to hit the key one extra
time within a 500 ms time window.

For arbitrary seek I didn't want to add any new seek keys so
I added a binding for *, which enters a safe arbitrary seek mode,
which you then follow with the same seek buttons as before. If
you don't enter this star character the old behavior is preserved.
i.e. if doesn't alias a channel it skips, otherwise it changes
the channel if in smart keys mode.

I also made it possible to select any teletext page, or use
any of the common ATSC subchannel separators, so long as the
entry begins with a number. So you can enter 5-1, if that is
how you named your channel, without needing to change the
audio select character or the subchannelsep character.



git-svn-id: http://svn.mythtv.org/trunk@7689 4b91aca1-2a1f-0410-a73a-a78f2df169e8