Fixes the smart startchannel setting.
It had broken due to a formatting change that removed a needed space
in the DB query, but I also wanted to improve it a little. It now 
tells you if there are no channels defined for the source, and even
tells you to check the console if there is a database error.

I also changed it so it does not give you the option of entering an 
invalid channel, while there is some convenience to this when setting
up a non-scannable source it is also highly prone to error. If there
are any objections to this please raise them in mythtv-dev, I'd like
to try to address any of problems with this in a less error-prone way.



git-svn-id: http://svn.mythtv.org/trunk@7738 4b91aca1-2a1f-0410-a73a-a78f2df169e8