Fixes #564.
The problem is really that keyframe distance detection is 
disabled when we use MARK_GOP_BYFRAME, but fixing that is
somewhat involved. Until I get to that, this two liner should
do just fine. It uses a keyframe distance of 18 for jumping
forward by a keyframe when the keyframedist looks suspect.
This should work for any keyframe distance over 9 and less than
19, i.e. the common MPEG-2 distances of 12,15,18 are all covered.



git-svn-id: http://svn.mythtv.org/trunk@7743 4b91aca1-2a1f-0410-a73a-a78f2df169e8