* Implement settings DB caching within MythContext.  We now cache the results   of MythContext::Get*Setting() and MythContext::Get*SettingOnHost() calls   to cut down on unnecessary database access.  Cached values are cleared   whenever any frontend goes into the mythfrontend setup screen and the   caching is turned OFF while in the settings screens.
  NOTE: This is not an excuse to forego caching within Myth classes, if a
        method uses a value from the DB settings table multiple times, the
        class should cache this value itself.

* Fix UIImageType to cache the PlayBoxTransparency setting internally so we
  don't have to re-get the value everytime an image is set such as on the
  status icons on the Watch Recordings screen when you scroll around through
  the recordings in the list.

The above two changes should make some things in Myth a bit snappier,
especially on the Watch Recordings screen.  Prior to this patch, scrolling
up/down through the list would cause about 10 database queries each time the
user scrolled.  Now that the settings are cached, the only DB querie
required when scrolling through recordings is to get the channel icon path.
Should be able to do something about that as well.  The speedup is more
noticeable if your database is remote to your frontend.

*******************************************************************************
* ANOTHER NOTE: This will require a "make clean" since it alters MythContext. *
*******************************************************************************



git-svn-id: http://svn.mythtv.org/trunk@7811 4b91aca1-2a1f-0410-a73a-a78f2df169e8