Updated dbcheck to initialize the database as it is now  rather than as it was and do a ton of updates.
 Removed the nestitle, neskeywords tables.

 Added a new lookup ability to pull Game info from the new RomDB
 database. This includes versioning of the information to allow
 for updates and resyncing later.

 Made gamescans smarter when they already have data to avoid unneeded
 work and speed up the scan when only small changes need to be done.

 Add publisher info and edit screens.

 Added a version flag to the metadata to easily determine what the
 user has customized, either for preservation or for export, or
 if it should be updated with updated RomDB information.
 Add more GameTypes.

 Modified a game scan to do a lot less SQL lookups descreasin overall
 the time it takes to complete a scan with metadata updates. the larger
 the collection the more of a savings.

=== Below is a copy of an email I am posting to mythtv-dev as well. ====


  Well for all of you MythGame users out there I just commited
some code that adds back in metadata gathering when you do a
game scan. However now there is data for a lot more gametypes
than before.

  Rather than use the old catver.ini, or the SQLite tosec database
or even the many dat files out there I decided to gather
everything to one large Rom Database  that I'm aptly named "RomDB"
After spending a couple months on and off looking at various sources
of info without finding any that  provided an "all in one"
solution that fit all of what I wanted I decided to basically
gather and keep my own.

 A new table called "romdb" has been created and I've included a copy of
the current romdb sql data in the contrib direcotry of the main
MythTV source which you will need to import if you want to take
advantage of this new meta data.   It's by no means complete
but it contains information on over 165k of roms and I intend to
keep filling in information as time goes on. I'll be including
a couple perl scripts later that can be used for updating the
romdb table without having to reimport all 165k entries as well
as for allowing users to pull out any custom rom information that
they filled in manually for those that want to submit that info
for inclusion in the main romdb list.

 I've put up a simple php script for doing web based searches against the
database for those who want to check it out at

 http://www.phaze.org/romdb/

 The name based searches are slow at the moment as they are full table scans
on a slow box but CRC searches are much faster. That site will also
always have a link to the latest SQL dump of the full romdb for anyone
to grab as they want.




git-svn-id: http://svn.mythtv.org/trunk@7822 4b91aca1-2a1f-0410-a73a-a78f2df169e8