Refs #340.
Simplify the TV class to use the TVChain-derived ProgramInfo for its OSD information, instead of using a query to the backend.  Store the channel name + input name in the TVChain table to make this easier.

Remove obsolete queries (GET_PROGRAM_INFO and GET_INPUT_NAME), and bump the protocol version to 21.


git-svn-id: http://svn.mythtv.org/trunk@7826 4b91aca1-2a1f-0410-a73a-a78f2df169e8