Fixes #542, hopefully without bad side-effects this time.
The problem with [7721] was the constraint change on line 118, which didn't do what I expected it too because of the type of qmat.

I've also kept one more parameter in a preloaded register this time in the H263 block, and got rid of the push/pop since it isn't needed with the REG_c reservation at the end of the asm blocks.



git-svn-id: http://svn.mythtv.org/trunk@7833 4b91aca1-2a1f-0410-a73a-a78f2df169e8