References #574. Don't free frame until we are done with it.
I thought this might be the cause for #574. But it is not.

My guess is that something is broken with seeking when using libmpeg2,
so that it seeks too far into an I frame, and misses parts of the image.

Does this happen with both PVR-250 and DVB/ATSC recordings?



git-svn-id: http://svn.mythtv.org/trunk@7834 4b91aca1-2a1f-0410-a73a-a78f2df169e8