Refs #340.
Seeking across file boundaries for livetv.  Possibly ff/rew too, didn't test yet.

Behavior is: When seeking forward, and not playing the most recent file, seek to the beginning of the next program.  When seeking backward, it'll go to 5 seconds before the end of the file.

There seem to be some issues syncing the position map for the chains.  Need to look into those later.


git-svn-id: http://svn.mythtv.org/trunk@7838 4b91aca1-2a1f-0410-a73a-a78f2df169e8